Better way of changing a user's role
