+++ BTCPayServer/Controllers/UIStoresController.cs
@@ -164,131 +164,8 @@ namespace BTCPayServer.Controllers
             }
             return Forbid();
         }
-
-        [HttpGet("{storeId}/users")]
-        public async Task<IActionResult> StoreUsers()
-        {
-            var vm = new StoreUsersViewModel { Role = StoreRoleId.Guest.Role };
-            await FillUsers(vm);
-            return View(vm);
-        }
-
-        private async Task FillUsers(StoreUsersViewModel vm)
-        {
-            var users = await _Repo.GetStoreUsers(CurrentStore.Id);
-            vm.StoreId = CurrentStore.Id;
-            vm.Users = users.Select(u => new StoreUsersViewModel.StoreUserViewModel()
-            {
-                Email = u.Email,
-                Id = u.Id,
-                Role = u.StoreRole.Role
-            }).ToList();
-        }
-
+        
         public StoreData? CurrentStore => HttpContext.GetStoreData();
-
-        [HttpPost("{storeId}/users")]
-        [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
-        public async Task<IActionResult> StoreUsers(string storeId, StoreUsersViewModel vm)
-        {
-            await FillUsers(vm);
-            if (!ModelState.IsValid)
-            {
-                return View(vm);
-            }
-
-            var roles = await _Repo.GetStoreRoles(CurrentStore.Id);
-            if (roles.All(role => role.Id != vm.Role))
-            {
-                ModelState.AddModelError(nameof(vm.Role), "Invalid role");
-                return View(vm);
-            }
-            
-            var user = await _UserManager.FindByEmailAsync(vm.Email);
-            var isExistingUser = user is not null;
-            var isExistingStoreUser = isExistingUser && await _Repo.GetStoreUser(storeId, user!.Id) is not null;
-            var successInfo = string.Empty;
-            if (user == null)
-            {
-                user = new ApplicationUser
-                {
-                    UserName = vm.Email,
-                    Email = vm.Email,
-                    RequiresEmailConfirmation = _policiesSettings.RequiresConfirmedEmail,
-                    RequiresApproval = _policiesSettings.RequiresUserApproval,
-                    Created = DateTimeOffset.UtcNow
-                };
-
-                var result = await _UserManager.CreateAsync(user);
-                if (result.Succeeded)
-                {
-                    var tcs = new TaskCompletionSource<Uri>();
-                    var currentUser = await _UserManager.GetUserAsync(HttpContext.User);
-
-                    _eventAggregator.Publish(new UserRegisteredEvent
-                    {
-                        RequestUri = Request.GetAbsoluteRootUri(),
-                        Kind = UserRegisteredEventKind.Invite,
-                        User = user,
-                        InvitedByUser = currentUser,
-                        CallbackUrlGenerated = tcs
-                    });
-                    
-                    var callbackUrl = await tcs.Task;
-                    var settings = await _settingsRepository.GetSettingAsync<EmailSettings>() ?? new EmailSettings();
-                    var info = settings.IsComplete()
-                        ? "An invitation email has been sent.<br/>You may alternatively"
-                        : "An invitation email has not been sent, because the server does not have an email server configured.<br/> You need to";
-                    successInfo = $"{info} share this link with them: <a class='alert-link' href='{callbackUrl}'>{callbackUrl}</a>";
-                }
-                else
-                {
-                    ModelState.AddModelError(nameof(vm.Email), "User could not be invited");
-                    return View(vm);
-                }
-            }
-
-            var roleId = await _Repo.ResolveStoreRoleId(storeId, vm.Role);
-            var action = isExistingUser
-                ? isExistingStoreUser ? "updated" : "added"
-                : "invited";
-            if (await _Repo.AddOrUpdateStoreUser(CurrentStore.Id, user.Id, roleId))
-            {
-                TempData.SetStatusMessageModel(new StatusMessageModel
-                {
-                    Severity = StatusMessageModel.StatusSeverity.Success,
-                    AllowDismiss = false,
-                    Html = $"User {action} successfully." + (string.IsNullOrEmpty(successInfo) ? "" : $" {successInfo}")
-                });
-                return RedirectToAction(nameof(StoreUsers));
-            }
-
-            ModelState.AddModelError(nameof(vm.Email), $"The user could not be {action}");
-            return View(vm);
-        }
-
-        [HttpGet("{storeId}/users/{userId}/delete")]
-        [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
-        public async Task<IActionResult> DeleteStoreUser(string userId)
-        {
-            var user = await _UserManager.FindByIdAsync(userId);
-            if (user == null)
-                return NotFound();
-            return View("Confirm", new ConfirmModel("Remove store user", $"This action will prevent <strong>{Html.Encode(user.Email)}</strong> from accessing this store and its settings. Are you sure?", "Remove"));
-        }
-
-        [HttpPost("{storeId}/users/{userId}/delete")]
-        [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
-        public async Task<IActionResult> DeleteStoreUserPost(string storeId, string userId)
-        {
-            if (await _Repo.RemoveStoreUser(storeId, userId))
-                TempData[WellKnownTempData.SuccessMessage] = "User removed successfully.";
-            else
-            {
-                TempData[WellKnownTempData.ErrorMessage] = "Removing this user would result in the store having no owner.";
-            }
-            return RedirectToAction(nameof(StoreUsers), new { storeId, userId });
-        }
 
         [HttpGet("{storeId}/rates")]
         public IActionResult Rates()
